refactor: improve components file structure
