+++ components/CheckConnectedWallet.tsx
