+++ components/index.ts
@@ -1,3 +1,10 @@
+export { default as Button } from './Button'
+export { default as CheckConnectedWallet } from './CheckConnectedWallet'
+export { default as Modal } from './Modal'
+export { default as Navbar } from './Navbar'
+export { default as Spinner } from './Spinner'
+export { default as Title } from './Title'
+export { default as Wallet } from './Wallet'
+
 export * from './cards'
 export * from './forms'
-export * from './layout'
